mise à jour des models et migrations. Ajout sauvegarde mensuel de la base de donnée

+++ database/migrations/2021_11_19_145349_create_sessions_table.php
@@ -17,7 +17,7 @@ class CreateSessionsTable extends Migration
         Schema::create('sessions', function (Blueprint $table) {
             $table->id();
             $table->string('name')->nullable();
-            $table->unsignedBigInteger('status')->default('0');
+            $table->integer('status')->default('0');
             $table->timestamps();
         });
     }
